% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca_clone_heterogeneity.R
\name{compute_pca_clone_heterogeneity}
\alias{compute_pca_clone_heterogeneity}
\title{Compute PCA-based clone heterogeneity (intra- and inter-clone distances)}
\usage{
compute_pca_clone_heterogeneity(
  seu,
  celltype_keep = c("LT-HSC", "ST-HSC"),
  clone_col = "CloneID_clean",
  sample_col = "sampleName",
  assay = "RNA",
  n_pcs = 30,
  min_pct = 0.1,
  min_mean = 0.1,
  min_cells = 2,
  group_recode = NULL,
  sample_order = NULL,
  prefix = "Clone Heterogeneity"
)
}
\description{
Compute PCA-based clone heterogeneity (intra- and inter-clone distances)
}
