% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variance_and_cv.R
\name{compute_variance_and_cv}
\alias{compute_variance_and_cv}
\title{Compute variance tests, CV, and mean-adjusted variance metrics}
\usage{
compute_variance_and_cv(
  seu,
  assay = "RNA",
  slot = "data",
  group_col = "donor_age",
  n_hvg = 2000,
  selection.method = "vst",
  n_cores = 1,
  chunk_size = 250,
  tests = c("brown_forsythe", "levene", "bartlett"),
  min_pct = 0.1,
  min_mean = 0.1
)
}
\description{
Compute variance tests, CV, and mean-adjusted variance metrics
}
