% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variance_metrics.R
\name{compute_variance_metrics}
\alias{compute_variance_metrics}
\title{Compute variance tests and mean-adjusted variance metrics}
\usage{
compute_variance_metrics(
  seu,
  assay = "RNA",
  layer = "data",
  group_col = "donor_age",
  n_hvg = 2000,
  selection.method = "vst",
  n_cores = 1,
  chunk_size = 250,
  min_pct = 0.1,
  min_mean = 0.1
)
}
\description{
Compute variance tests and mean-adjusted variance metrics
}
