% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_heterogeneity.R
\name{plot_density_gene}
\alias{plot_density_gene}
\title{Plot gene expression density between two groups}
\usage{
plot_density_gene(
  seurat_obj,
  gene,
  group_col,
  group1_labels,
  group2_labels,
  group1_name = "Group1",
  group2_name = "Group2",
  title_prefix = "",
  dvgs_df = NULL,
  mean_adj_var_group1_col = NULL,
  mean_adj_var_group2_col = NULL,
  group_colors = c("#89AEEB", "#F5A36C")
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{gene}{Gene name}

\item{group_col}{Metadata column defining groups}

\item{group1_labels}{Labels belonging to group1}

\item{group2_labels}{Labels belonging to group2}

\item{group1_name}{Display name of group1}

\item{group2_name}{Display name of group2}

\item{title_prefix}{Plot title prefix}

\item{dvgs_df}{Dataframe containing mean-adjusted variance}

\item{mean_adj_var_group1_col}{Column for group1 variance}

\item{mean_adj_var_group2_col}{Column for group2 variance}

\item{group_colors}{Color vector for the two groups}
}
\value{
ggplot object
}
\description{
Visualizes gene expression distributions and optional mean-adjusted variance
between two groups of cells in a Seurat object.
}
