% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variance_cv_summaries.R
\name{plot_variance_cv_summaries}
\alias{plot_variance_cv_summaries}
\title{Plot variance, CV, and mean-adjusted variance summaries}
\usage{
plot_variance_cv_summaries(
  df,
  fdr_col = "fdr_brown_forsythe",
  fc_col = "log2FC_mean_adjusted_variance",
  fc_col_var = "log2FC_variance",
  fdr_cutoff = 0.05,
  fc_cutoff = 0.25,
  color_palette = c("#2F7BAA", "#E58B1C"),
  prefix = "Day0",
  group_order = NULL
)
}
\description{
Plot variance, CV, and mean-adjusted variance summaries
}
