% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_clone_distribution_engine.R
\name{run_clone_distribution_engine}
\alias{run_clone_distribution_engine}
\title{Wrapper function that produce distribution plots and statistical summary}
\usage{
run_clone_distribution_engine(
  comparison_list,
  output_file = NULL,
  rep_col = "Rep",
  clone_col = "CloneID"
)
}
\description{
Wrapper function that produce distribution plots and statistical summary
}
