% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_low_high_OA_analysis_for_a_single_sample.R
\name{run_low_high_OA_analysis_for_a_single_sample}
\alias{run_low_high_OA_analysis_for_a_single_sample}
\title{OA function (per-sample input)}
\usage{
run_low_high_OA_analysis_for_a_single_sample(
  seurat_obj,
  sample_label = "Sample",
  output_dir = "OA_Analysis",
  clone_col = "CloneID",
  HSC_types = c("LT-HSC", "ST-HSC"),
  nonHSC_types = NULL,
  deg_celltypes = c("LT-HSC", "ST-HSC", "MPP"),
  top_hvgs = 2000,
  min_detect_prop = 0.1,
  min_mean_norm = 0.1,
  min_mean_raw = 2,
  fc_cutoff = 0.25,
  fdr_cutoff = 0.05,
  low_output_markers,
  high_output_markers,
  celltype_colors,
  full_ref_deg_genes_list
)
}
\description{
OA function (per-sample input)
}
