% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scCloneVar_test_demo.R
\docType{data}
\name{scCloneVar_test_demo}
\alias{scCloneVar_test_demo}
\title{scCloneVar_test_demo}
\format{
A Seurat object
}
\source{
Internal subset of in vitro cross-age dataset
}
\usage{
scCloneVar_test_demo
}
\description{
A small Seurat object containing 20 Young and 20 Old clones
derived from the in vitro cross-age dataset.
}
\details{
The dataset includes:
\itemize{
\item RNA assay
\item PCA and UMAP reductions
\item CloneID and donor_age metadata
}

This dataset is intended for demonstrating scCloneVar workflows.
}
\keyword{datasets}
